tg bot bitcoin

+++ src/main/java/com/skillbox/cryptobot/utils/TextUtil.java
@@ -1,8 +1,10 @@
 package com.skillbox.cryptobot.utils;
 
+import java.math.BigDecimal;
+
 public class TextUtil {
 
-    public static String toString(double value) {
+    public static String toString(BigDecimal value) {
         return String.format("%.3f", value);
     }
 }
